pyproject.toml
[build-system]
requires = ["maturin>=0.14,<0.15"]
build-backend = "maturin"

[project]
name = "resvg-py"
version = "0.0.0"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "numpy",
    "Pillow",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-snapshot"]

[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module"]

[tool.ruff]
select = ["ALL"]
target-version = "py37"

[tool.ruff.isort]
known-first-party = ["resvg_py"]

[tool.ruff.per-file-ignores]
"**/*.pyi" = ["D1"]
"tests/test_*.py" = ["D1"]
